Climate strict handling of current temp value
Measured value of room temperature

This property indicates the measured room temperature (°C). The property value range shall be 0x81 to 0x7D (-127 to 125°C). When the property value of the actual piece of equipment is higher than the upper limit of the property value range, the overflow code 0x7F shall be used. When the property value is lower than the lower limit of the property value range, the underflow code 0x80 shall be used. When the measurement value cannot be returned, 0x7E shall be used.